separate individual/company packages + move licence package under hecTaxCheck
