+++ app/uk/gov/hmrc/hec/models/hecTaxCheck/licence/LicenceType.scala
@@ -14,29 +14,34 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.hec.models.licence
+package uk.gov.hmrc.hec.models.hecTaxCheck.licence
 
 import ai.x.play.json.Jsonx
 import ai.x.play.json.SingletonEncoder.simpleName
 import ai.x.play.json.implicits.formatSingleton
-import enumeratum.{Enum, EnumEntry}
+import cats.Eq
+import enumeratum._
 import play.api.libs.json.Format
 
 import scala.collection.immutable
 
-sealed trait LicenceValidityPeriod extends EnumEntry with Product with Serializable
+sealed trait LicenceType extends EnumEntry with Product with Serializable
 
-object LicenceValidityPeriod extends Enum[LicenceValidityPeriod] {
+object LicenceType extends Enum[LicenceType] {
 
-  case object UpToOneYear extends LicenceValidityPeriod
-  case object UpToTwoYears extends LicenceValidityPeriod
-  case object UpToThreeYears extends LicenceValidityPeriod
-  case object UpToFourYears extends LicenceValidityPeriod
-  case object UpToFiveYears extends LicenceValidityPeriod
+  case object DriverOfTaxisAndPrivateHires extends LicenceType
 
-  val values: immutable.IndexedSeq[LicenceValidityPeriod] = findValues
+  case object OperatorOfPrivateHireVehicles extends LicenceType
+
+  case object ScrapMetalMobileCollector extends LicenceType
+
+  case object ScrapMetalDealerSite extends LicenceType
+
+  val values: immutable.IndexedSeq[LicenceType] = findValues
 
   @SuppressWarnings(Array("org.wartremover.warts.Throw", "org.wartremover.warts.Equals"))
-  implicit val format: Format[LicenceValidityPeriod] = Jsonx.formatSealed[LicenceValidityPeriod]
+  implicit val format: Format[LicenceType] = Jsonx.formatSealed[LicenceType]
+
+  implicit val eq: Eq[LicenceType] = Eq.fromUniversalEquals
 
 }
