+++ app/uk/gov/hmrc/hec/models/hecTaxCheck/licence/LicenceValidityPeriod.scala
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.hec.models.hecTaxCheck.licence
+
+import ai.x.play.json.Jsonx
+import ai.x.play.json.SingletonEncoder.simpleName
+import ai.x.play.json.implicits.formatSingleton
+import enumeratum.{Enum, EnumEntry}
+import play.api.libs.json.Format
+
+import scala.collection.immutable
+
+sealed trait LicenceValidityPeriod extends EnumEntry with Product with Serializable
+
+object LicenceValidityPeriod extends Enum[LicenceValidityPeriod] {
+
+  case object UpToOneYear extends LicenceValidityPeriod
+  case object UpToTwoYears extends LicenceValidityPeriod
+  case object UpToThreeYears extends LicenceValidityPeriod
+  case object UpToFourYears extends LicenceValidityPeriod
+  case object UpToFiveYears extends LicenceValidityPeriod
+
+  val values: immutable.IndexedSeq[LicenceValidityPeriod] = findValues
+
+  @SuppressWarnings(Array("org.wartremover.warts.Throw", "org.wartremover.warts.Equals"))
+  implicit val format: Format[LicenceValidityPeriod] = Jsonx.formatSealed[LicenceValidityPeriod]
+
+}
