+++ app/uk/gov/hmrc/hec/models/taxCheckMatch/HECTaxCheckMatchRequest.scala
@@ -17,9 +17,10 @@
 package uk.gov.hmrc.hec.models.taxCheckMatch
 
 import play.api.libs.json.{Json, OFormat}
-import uk.gov.hmrc.hec.models.hecTaxCheck.{DateOfBirth, HECTaxCheckCode}
+import uk.gov.hmrc.hec.models.hecTaxCheck.HECTaxCheckCode
+import uk.gov.hmrc.hec.models.hecTaxCheck.individual.DateOfBirth
 import uk.gov.hmrc.hec.models.ids.CRN
-import uk.gov.hmrc.hec.models.licence.LicenceType
+import uk.gov.hmrc.hec.models.hecTaxCheck.licence.LicenceType
 import uk.gov.hmrc.hec.util.EitherUtils.eitherFormat
 
 final case class HECTaxCheckMatchRequest(
