+++ test/uk/gov/hmrc/hec/services/FileCreationServiceSpec.scala
@@ -24,10 +24,12 @@ import uk.gov.hmrc.hec.models.hecTaxCheck.ApplicantDetails.{CompanyApplicantDeta
 import uk.gov.hmrc.hec.models.hecTaxCheck.HECTaxCheckData.{CompanyHECTaxCheckData, IndividualHECTaxCheckData}
 import uk.gov.hmrc.hec.models.hecTaxCheck.HECTaxCheckSource.Digital
 import uk.gov.hmrc.hec.models.hecTaxCheck.TaxDetails.{CompanyTaxDetails, IndividualTaxDetails}
-import uk.gov.hmrc.hec.models.ids.{CRN, CTUTR, GGCredId, NINO, SAUTR}
-import uk.gov.hmrc.hec.models.licence.LicenceType.{DriverOfTaxisAndPrivateHires, OperatorOfPrivateHireVehicles, ScrapMetalDealerSite, ScrapMetalMobileCollector}
-import uk.gov.hmrc.hec.models.licence.{LicenceTimeTrading, LicenceType, LicenceValidityPeriod}
-import uk.gov.hmrc.hec.models.hecTaxCheck.{CTAccountingPeriod, CTStatus, CTStatusResponse, CompanyHouseName, CorrectiveAction, DateOfBirth, HECTaxCheck, HECTaxCheckCode, HECTaxCheckFileBodyList, LicenceDetails, Name, SAStatus, SAStatusResponse, TaxSituation, TaxYear, YesNoAnswer}
+import uk.gov.hmrc.hec.models.hecTaxCheck.company.{CTAccountingPeriod, CTStatus, CTStatusResponse, CompanyHouseName}
+import uk.gov.hmrc.hec.models.hecTaxCheck.individual.{DateOfBirth, Name, SAStatus, SAStatusResponse}
+import uk.gov.hmrc.hec.models.hecTaxCheck.licence.LicenceType.{DriverOfTaxisAndPrivateHires, OperatorOfPrivateHireVehicles, ScrapMetalDealerSite, ScrapMetalMobileCollector}
+import uk.gov.hmrc.hec.models.hecTaxCheck.licence.{LicenceDetails, LicenceTimeTrading, LicenceType, LicenceValidityPeriod}
+import uk.gov.hmrc.hec.models.hecTaxCheck._
+import uk.gov.hmrc.hec.models.ids._
 import uk.gov.hmrc.hec.models.{Error, hecTaxCheck}
 import uk.gov.hmrc.hec.util.TimeProvider
 
@@ -191,7 +193,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.FourToEightYears,
                   LicenceValidityPeriod.UpToFiveYears,
                   TaxSituation.PAYE,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.ReturnFound)),
+                  Some(SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.ReturnFound)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 ),
                 createIndividualHecTaxCheck(
@@ -199,7 +201,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.TwoToFourYears,
                   LicenceValidityPeriod.UpToFiveYears,
                   TaxSituation.PAYE,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoticeToFileIssued)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoticeToFileIssued)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 ),
                 createIndividualHecTaxCheck(
@@ -207,7 +209,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.TwoToFourYears,
                   LicenceValidityPeriod.UpToFiveYears,
                   TaxSituation.PAYE,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoReturnFound)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoReturnFound)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 )
               )
@@ -241,7 +243,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.FourToEightYears,
                   LicenceValidityPeriod.UpToOneYear,
                   TaxSituation.SA,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.ReturnFound)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.ReturnFound)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 ),
                 createIndividualHecTaxCheck(
@@ -249,7 +251,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.TwoToFourYears,
                   LicenceValidityPeriod.UpToTwoYears,
                   TaxSituation.SA,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoticeToFileIssued)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoticeToFileIssued)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 ),
                 createIndividualHecTaxCheck(
@@ -257,7 +259,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.TwoToFourYears,
                   LicenceValidityPeriod.UpToThreeYears,
                   TaxSituation.SA,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoReturnFound)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoReturnFound)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 )
               )
@@ -292,7 +294,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.FourToEightYears,
                   LicenceValidityPeriod.UpToOneYear,
                   TaxSituation.SAPAYE,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.ReturnFound)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.ReturnFound)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 ),
                 createIndividualHecTaxCheck(
@@ -300,7 +302,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.TwoToFourYears,
                   LicenceValidityPeriod.UpToTwoYears,
                   TaxSituation.SAPAYE,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoticeToFileIssued)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoticeToFileIssued)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 ),
                 createIndividualHecTaxCheck(
@@ -308,7 +310,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.TwoToFourYears,
                   LicenceValidityPeriod.UpToThreeYears,
                   TaxSituation.SAPAYE,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoReturnFound)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoReturnFound)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 )
               )
@@ -343,7 +345,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.FourToEightYears,
                   LicenceValidityPeriod.UpToOneYear,
                   TaxSituation.NotChargeable,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.ReturnFound)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.ReturnFound)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 ),
                 createIndividualHecTaxCheck(
@@ -351,7 +353,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.TwoToFourYears,
                   LicenceValidityPeriod.UpToTwoYears,
                   TaxSituation.NotChargeable,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoticeToFileIssued)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoticeToFileIssued)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 ),
                 createIndividualHecTaxCheck(
@@ -359,7 +361,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
                   LicenceTimeTrading.TwoToFourYears,
                   LicenceValidityPeriod.UpToThreeYears,
                   TaxSituation.NotChargeable,
-                  Some(hecTaxCheck.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoReturnFound)),
+                  Some(individual.SAStatusResponse(SAUTR("1234567"), TaxYear(2021), SAStatus.NoReturnFound)),
                   Some(CorrectiveAction.RegisterNewSAAccount)
                 )
               )
@@ -433,7 +435,7 @@ class FileCreationServiceSpec extends AnyWordSpec with Matchers with MockFactory
               hecTaxCheck.HECTaxCheck(
                 CompanyHECTaxCheckData(
                   companyDetails,
-                  hecTaxCheck.LicenceDetails(licenceType, licenceTimeTrading, licenceValidityPeriod),
+                  LicenceDetails(licenceType, licenceTimeTrading, licenceValidityPeriod),
                   createTaxDetails(
                     ctStatusResponse,
                     ctIncomeDeclared,
